# docs/swagger.yaml
basePath: /
definitions:
  internal_infrastructure_httpserver.CreateTenantRequest:
    properties:
      expires_in:
        description: ExpiresIn is the JWT lifetime in seconds. Defaults to 86400 (24
          h). Only applies in jwt mode.
        example: 86400
        type: integer
      roles:
        description: Roles assigned to the credential. Defaults to ["writer","reader"]
          if omitted.
        example:
        - writer
        - reader
        items:
          type: string
        type: array
      subject_id:
        description: SubjectID identifies the principal within the tenant (e.g. "admin",
          "service-account").
        example: acme-admin
        type: string
      tenant_id:
        description: TenantID is the unique identifier for the tenant (e.g. "acme-corp").
        example: acme-corp
        type: string
    type: object
  internal_infrastructure_httpserver.ErrorResponse:
    properties:
      error:
        example: stream_id is required
        type: string
    type: object
  internal_infrastructure_httpserver.EventResponse:
    properties:
      id:
        description: ID is the globally unique event UUID.
        example: 01906c2e-4a3b-7000-8000-abc123def456
        type: string
      metadata:
        additionalProperties:
          type: string
        description: Metadata is the key-value metadata attached to the event.
        type: object
      occurred_at:
        description: OccurredAt is the UTC timestamp when the event occurred.
        example: "2026-04-21T10:00:00Z"
        type: string
      payload:
        description: Payload is the arbitrary JSON payload.
        type: object
      source:
        description: Source is the service that produced the event.
        example: orders-svc
        type: string
      stream_id:
        description: StreamID is the logical stream this event belongs to.
        example: order:1
        type: string
      tenant_id:
        description: TenantID scopes the event to a tenant; derived from the caller's
          Identity.
        example: default
        type: string
      type:
        description: Type is the event type in dot-notation.
        example: order.created
        type: string
      version:
        description: Version is monotonically increasing within the stream, assigned
          by the store.
        example: 1
        type: integer
    type: object
  internal_infrastructure_httpserver.HealthResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  internal_infrastructure_httpserver.IngestRequest:
    properties:
      metadata:
        additionalProperties:
          type: string
        description: Metadata is optional key-value metadata for routing or tracing.
        example:
          region: us-east-1
          trace_id: abc-123
        type: object
      payload:
        description: Payload is an arbitrary JSON object attached to the event.
        type: object
      source:
        description: Source is the originating service name.
        example: orders-svc
        type: string
      stream_id:
        description: StreamID is the target stream. Created implicitly if it does
          not exist.
        example: order:1
        type: string
      type:
        description: Type is the event type in dot-notation.
        example: order.created
        type: string
    type: object
  internal_infrastructure_httpserver.IssueTokenRequest:
    properties:
      expires_in:
        description: ExpiresIn is the desired token lifetime in seconds. Defaults
          to 3600 (1 hour). Max 86400 (24 h).
        example: 3600
        type: integer
    type: object
  internal_infrastructure_httpserver.IssuedTokenResponse:
    properties:
      expires_at:
        description: ExpiresAt is the UTC timestamp when the token expires.
        example: "2026-04-21T11:00:00Z"
        type: string
      subject_id:
        description: SubjectID identifies the principal this token was issued for.
        example: api-key
        type: string
      tenant_id:
        description: TenantID is the tenant scope embedded in the token.
        example: default
        type: string
      token:
        description: 'Token is the signed JWT to use in subsequent requests as "Authorization:
          Bearer <token>".'
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  internal_infrastructure_httpserver.ListEventsResponse:
    properties:
      events:
        description: Events is the page of events sorted by occurred_at DESC.
        items:
          $ref: '#/definitions/internal_infrastructure_httpserver.EventResponse'
        type: array
      limit:
        description: Limit is the page size applied to this query.
        example: 20
        type: integer
      offset:
        description: Offset is the number of events skipped.
        example: 0
        type: integer
      read_model:
        description: ReadModel identifies the data source ("elasticsearch").
        example: elasticsearch
        type: string
      total:
        description: Total is the total number of events in the index (for pagination).
        example: 128
        type: integer
    type: object
  internal_infrastructure_httpserver.QueryResultResponse:
    properties:
      events:
        description: Events is the ordered (version ASC) page of events.
        items:
          $ref: '#/definitions/internal_infrastructure_httpserver.EventResponse'
        type: array
      limit:
        description: Limit is the page size applied to this query.
        example: 20
        type: integer
      offset:
        description: Offset is the number of events skipped.
        example: 0
        type: integer
      read_model:
        description: ReadModel identifies the data source ("elasticsearch").
        example: elasticsearch
        type: string
      stream_id:
        description: StreamID is the queried stream identifier.
        example: order:1
        type: string
      total:
        description: Total is the total number of matching events, useful for pagination.
        example: 42
        type: integer
    type: object
  internal_infrastructure_httpserver.TenantCredentialsResponse:
    properties:
      expires_at:
        description: ExpiresAt is the UTC expiry of the issued token. Only present
          in jwt mode.
        example: "2026-04-22T21:57:00Z"
        type: string
      mode:
        description: Mode reflects the active auth mode ("jwt" or "simple").
        example: jwt
        type: string
      note:
        description: Note is a human-readable hint returned in simple mode.
        example: 'simple mode is single-tenant — use X-API-Key: <AUTH_API_KEY>'
        type: string
      subject_id:
        description: SubjectID is the principal this credential belongs to.
        example: acme-admin
        type: string
      tenant_id:
        description: TenantID is the provisioned tenant identifier.
        example: acme-corp
        type: string
      token:
        description: 'Token is the signed JWT to use as "Authorization: Bearer <token>".
          Only present in jwt mode.'
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  internal_infrastructure_httpserver.TokenResponse:
    properties:
      roles:
        description: Roles lists the permissions granted to this credential.
        example:
        - writer
        - reader
        items:
          type: string
        type: array
      subject_id:
        description: SubjectID identifies the authenticated principal.
        example: api-key
        type: string
      tenant_id:
        description: TenantID is the tenant scope associated with this credential.
        example: default
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    HTTP API for ingesting and querying events in the Event Streaming & Audit Platform.

    ## Authentication
    All /events endpoints require authentication. /health and /swagger are public.
    Set AUTH_MODE=simple (default) to use X-API-Key, or AUTH_MODE=jwt for Bearer tokens.

    ## Data consistency
    Write path: PostgreSQL (source of truth) → Kafka → Elasticsearch (read model).
    GET /events/{streamID} is served from Elasticsearch and is eventually consistent.
  title: Event Streaming and Audit API
  version: "1.0"
paths:
  /auth/issue:
    post:
      consumes:
      - application/json
      description: |-
        Authenticates the caller using their current credentials (API key or existing JWT)
        and mints a new signed JWT that can be used as a Bearer token.

        The issued token inherits the caller's **subject_id**, **tenant_id**, and **roles**,
        so it carries exactly the same authorization scope as the originating credential.

        **Requires** `AUTH_JWT_SECRET` to be configured on the server — returns 501 otherwise.

        Typical use-cases:
        - Upgrade from static API key to short-lived JWT
        - Refresh an expiring JWT before it expires
        - Issue scoped tokens for downstream services
      parameters:
      - description: Optional — override token lifetime
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_infrastructure_httpserver.IssueTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.IssuedTokenResponse'
        "401":
          description: Missing or invalid credentials
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
        "500":
          description: Failed to sign token
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
        "501":
          description: Token issuance not configured (AUTH_JWT_SECRET missing)
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Issue a JWT access token
      tags:
      - auth
  /auth/token:
    post:
      description: |-
        Validates the credentials supplied in the request headers and returns the
        caller's Identity (subject, tenant_id, roles). No request body required.

        **Simple mode** (AUTH_MODE=simple): send `X-API-Key: <key>` header.
        **JWT mode** (AUTH_MODE=jwt): send `Authorization: Bearer <token>` header.

        Use this endpoint to verify that credentials are accepted before making
        other API calls, and to discover the tenant_id scoped to your key/token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.TokenResponse'
        "401":
          description: Missing or invalid credentials
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Validate credentials and retrieve identity
      tags:
      - auth
  /events:
    get:
      description: |-
        Returns a paginated list of events across all streams, sorted by
        occurred_at DESC (newest first). Useful as an audit feed or event log.

        Results come from the Elasticsearch read model and are **eventually consistent**
        with the PostgreSQL event store — recently ingested events may not appear immediately.
      parameters:
      - description: Page size — default 20, max 100
        in: query
        name: limit
        type: integer
      - description: Number of events to skip — default 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ListEventsResponse'
        "401":
          description: Missing or invalid credentials
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
        "500":
          description: Read model query failed
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List all events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: |-
        Appends a new event to the specified stream. The event is persisted in PostgreSQL
        synchronously, then published to Kafka best-effort. A Kafka failure does NOT fail
        this request — the event is already durable in the store. Version is assigned by
        the store (MAX+1 per stream) and returned in the response.
      parameters:
      - description: Event payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_infrastructure_httpserver.IngestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.EventResponse'
        "400":
          description: Invalid JSON body
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
        "401":
          description: Missing or invalid credentials
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
        "422":
          description: Validation error (missing required fields)
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Ingest an event
      tags:
      - events
  /events/{streamID}:
    get:
      description: |-
        Returns a paginated, version-ordered list of events for a stream from the
        Elasticsearch read model. Results are **eventually consistent** with the
        PostgreSQL event store — recently ingested events may not be visible immediately.
        Response headers communicate the data source and consistency level:
        - X-Read-Model: elasticsearch
        - X-Data-Consistency: eventual
      parameters:
      - description: Stream identifier (e.g. order:1, user:42)
        in: path
        name: streamID
        required: true
        type: string
      - description: Page size — default 20, max 100
        in: query
        name: limit
        type: integer
      - description: Number of events to skip — default 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.QueryResultResponse'
        "400":
          description: stream_id missing
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
        "401":
          description: Missing or invalid credentials
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
        "500":
          description: Read model query failed
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Query events by stream
      tags:
      - events
  /health:
    get:
      description: Returns "ok" if the service is running. No authentication required.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.HealthResponse'
      summary: Health check
      tags:
      - health
  /tenants:
    post:
      consumes:
      - application/json
      description: |-
        Creates a logical tenant identity and returns the credentials needed to start
        making API calls. This is the **bootstrap endpoint** — use it once per tenant
        to obtain the first token or API key.

        Protected by `X-Admin-Key` header (configured via `ADMIN_KEY` env var).
        Keep this key secret and separate from regular API credentials.

        **JWT mode** (`AUTH_MODE=jwt`): issues a signed JWT embedding tenant_id, subject_id,
        and roles. Use the returned `token` as `Authorization: Bearer <token>`.

        **Simple mode** (`AUTH_MODE=simple`): single-tenant only — returns the static
        `AUTH_API_KEY` hint since all callers share the same credential.
      parameters:
      - description: Admin secret key (ADMIN_KEY env var)
        in: header
        name: X-Admin-Key
        required: true
        type: string
      - description: Tenant identity and credential options
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_infrastructure_httpserver.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.TenantCredentialsResponse'
        "400":
          description: Missing required fields
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
        "401":
          description: Invalid or missing X-Admin-Key
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
        "500":
          description: Failed to issue token
          schema:
            $ref: '#/definitions/internal_infrastructure_httpserver.ErrorResponse'
      summary: Provision a new tenant and issue initial credentials
      tags:
      - tenants
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'Static API key. Default value for local dev: dev-api-key. Configured
      via AUTH_API_KEY env var.'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}". Enable with AUTH_MODE=jwt
      and AUTH_JWT_SECRET env var.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"